+++ apps/fluent-tester/src/FluentTester/TestComponents/Menu/MenuTest.tsx
@@ -1,8 +1,16 @@
 import * as React from 'react';
 import { ButtonV1 as Button } from '@fluentui/react-native';
-import { Menu, MenuItem, MenuItemCheckbox, MenuTrigger, MenuPopover, MenuList, MenuDivider } from '@fluentui-react-native/menu';
+import {
+  Menu,
+  MenuItem,
+  MenuItemCheckbox,
+  MenuItemRadio,
+  MenuTrigger,
+  MenuPopover,
+  MenuList,
+  MenuDivider,
+} from '@fluentui-react-native/menu';
 import { Stack } from '@fluentui-react-native/stack';
-import { Text } from '@fluentui-react-native/experimental-text';
 import { stackStyle } from '../Common/styles';
 import { MENU_TESTPAGE } from './consts';
 import { Test, TestSection, PlatformStatus } from '../Test';
@@ -16,7 +24,9 @@ const MenuDefault: React.FunctionComponent = () => {
         </MenuTrigger>
         <MenuPopover>
           <MenuList>
-            <Text>Hello world!!!</Text>
+            <MenuItem content="A plain MenuItem" />
+            <MenuItem disabled content="A second disabled plain MenuItem" />
+            <MenuItem content="A third plain MenuItem" />
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -47,7 +57,8 @@ const MenuCheckmarks: React.FunctionComponent = () => {
           <MenuList>
             <MenuItem content="A plain MenuItem" />
             <MenuItemCheckbox name="itemTwo" content="A MenuItem with checkmark" />
-            <MenuItemCheckbox name="itemThree" content="A MenuItem with checkmark" />
+            <MenuItemCheckbox disabled name="itemThree" content="A disabled MenuItem with checkmark" />
+            <MenuItemCheckbox name="itemFour" content="A MenuItem with checkmark" />
           </MenuList>
         </MenuPopover>
       </Menu>
@@ -67,6 +78,25 @@ const MenuCheckmarks: React.FunctionComponent = () => {
   );
 };
 
+const MenuRadioItem: React.FunctionComponent = () => {
+  return (
+    <Stack style={stackStyle}>
+      <Menu defaultChecked={{ itemOne: true }}>
+        <MenuTrigger>
+          <Button>Items with radio selection</Button>
+        </MenuTrigger>
+        <MenuPopover>
+          <MenuList>
+            <MenuItemRadio name="itemOne" content="A MenuItem with checkmark" />
+            <MenuItemRadio name="itemTwo" content="Another MenuItem with checkmark" />
+            <MenuItemRadio name="itemThree" content="A third MenuItem with checkmark" />
+          </MenuList>
+        </MenuPopover>
+      </Menu>
+    </Stack>
+  );
+};
+
 const Submenu: React.FunctionComponent = () => {
   return (
     <Menu>
@@ -75,7 +105,8 @@ const Submenu: React.FunctionComponent = () => {
       </MenuTrigger>
       <MenuPopover>
         <MenuList>
-          <Text>Hello world!!!</Text>
+          <MenuItem content="A nested MenuItem" />
+          <MenuItem content="A second nested MenuItem" />
         </MenuList>
       </MenuPopover>
     </Menu>
@@ -147,6 +178,10 @@ const menuSections: TestSection[] = [
   {
     name: 'Menu Checkmarks',
     component: MenuCheckmarks,
+  },
+  {
+    name: 'Menu Radioitem',
+    component: MenuRadioItem,
   },
   {
     name: 'Menu open on hover',