+++ packages/experimental/Menu/src/Menu/useMenu.ts
@@ -10,7 +10,13 @@ export const useMenu = (props: MenuProps): MenuState => {
   const isControlled = typeof props.open !== 'undefined';
   const [open, setOpen] = useMenuOpenState(isControlled, props);
 
+  // Default behaviot for submenu is to open on hover
+  // the ...props line below will override this behavior for a submenu
+  // or apply openOnHover if passed into a root Menu.
+  const openOnHover = isSubmenu;
+
   return {
+    openOnHover,
     ...props,
     open,
     setOpen,