+++ packages/experimental/Menu/src/MenuPopover/MenuPopover.types.ts
@@ -1,12 +1,15 @@
-import type { IViewProps } from '@fluentui-react-native/adapters';
-import { DismissBehaviors } from '@fluentui-react-native/callout';
+import { DirectionalHint, DismissBehaviors, ICalloutProps } from '@fluentui-react-native/callout';
+import { AccessibilityRole } from 'react-native';
 
 export const menuPopoverName = 'MenuPopover';
 
-export interface MenuPopoverProps extends Omit<IViewProps, 'onPress'> {}
+export type MenuPopoverProps = ICalloutProps;
 
 export interface MenuPopoverState {
+  accessibilityRole: AccessibilityRole;
+  directionalHint?: DirectionalHint;
   dismissBehaviors: DismissBehaviors[];
+  doNotTakePointerCapture: boolean;
   onDismiss: () => void;
   setInitialFocus: boolean;
   triggerRef: React.RefObject<React.Component>;