+++ packages/experimental/Menu/src/MenuPopover/useMenuPopover.ts
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Platform } from 'react-native';
-import { DismissBehaviors } from '@fluentui-react-native/callout';
+import { I18nManager, Platform } from 'react-native';
+import { DirectionalHint, DismissBehaviors } from '@fluentui-react-native/callout';
 import { useMenuContext } from '../context/menuContext';
 import { MenuPopoverProps, MenuPopoverState } from './MenuPopover.types';
 
@@ -13,10 +13,25 @@ export const useMenuPopover = (_props: MenuPopoverProps): MenuPopoverState => {
   const dismissBehaviors = context.isControlled
     ? (['preventDismissOnKeyDown', 'preventDismissOnClickOutside'] as DismissBehaviors[])
     : undefined;
+  const directionalHint = getDirectionalHint(context.isSubmenu, I18nManager.isRTL);
 
   // Initial focus behavior differs per platform, Windows platforms move focus
   // automatically onto first element of Callout
   const setInitialFocus = Platform.OS === ('win32' as any) || Platform.OS === 'windows';
+  const doNotTakePointerCapture = context.openOnHover;
+  const accessibilityRole = 'menu';
 
-  return { triggerRef, onDismiss, dismissBehaviors, setInitialFocus };
+  return { accessibilityRole, triggerRef, onDismiss, directionalHint, dismissBehaviors, doNotTakePointerCapture, setInitialFocus };
+};
+
+const getDirectionalHint = (isSubmenu: boolean, isRtl: boolean): DirectionalHint | undefined => {
+  if (!isSubmenu) {
+    return undefined;
+  }
+
+  if (isRtl) {
+    return 'leftTopEdge';
+  }
+
+  return 'rightTopEdge';
 };