+++ packages/experimental/Tabs/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/experimental-tabs
 
-This log was last generated on Sat, 21 May 2022 16:18:06 GMT and should not be manually modified.
+This log was last generated on Thu, 26 May 2022 21:22:22 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.49
+
+Thu, 26 May 2022 21:22:22 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.16.1
+- Bump @fluentui-react-native/focus-zone to v0.9.30
+- Bump @fluentui-react-native/icon to v0.12.0
+
+## 0.4.48
+
+Thu, 26 May 2022 00:33:11 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.16.0
+- Bump @fluentui-react-native/focus-zone to v0.9.29
+
+## 0.4.47
+
+Wed, 25 May 2022 18:43:08 GMT
+
+### Patches
+
+- beachball sync (krsiler@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.7.30
+- Bump @fluentui-react-native/interactive-hooks to v0.15.10
+- Bump @fluentui-react-native/focus-zone to v0.9.28
+- Bump @fluentui-react-native/experimental-text to v0.9.0
+- Bump @fluentui-react-native/tokens to v0.14.0
+- Bump @fluentui-react-native/icon to v0.11.25
 
 ## 0.4.46